Kubernetes/05-SportIBook/backend/requirements.txt
Django>=2.0,<3.0
mysqlclient
djangorestframework
Pillow
uwsgi
dj-static
